internal/dag: Move dag.Builder data into a new struct
Break out data components from the dag.Builder into its own struct
to seperate the builder cache data apart.

Updates #2226
